kubernetes/src/controller/requirements_dev.txt
kopf[uvloop]~=1.43.0
pyyaml~=6.0.3
datamodel-code-generator~=0.54.0